+++ analyzers/tests/SonarAnalyzer.UnitTest/AnalysisContext/SonarAnalysisContextTest.ShouldAnalyze.cs
@@ -20,7 +20,7 @@
 
 using SonarAnalyzer.Common;
 
-namespace SonarAnalyzer.UnitTest.Helpers;
+namespace SonarAnalyzer.UnitTest.AnalysisContext;
 
 public partial class SonarAnalysisContextTest
 {
