+++ analyzers/tests/SonarAnalyzer.UnitTest/AnalysisContext/SonarAnalysisContextTest.ShouldAnalyzeGenerated.cs
@@ -23,7 +23,7 @@ using Microsoft.CodeAnalysis.Text;
 using Moq;
 using SonarAnalyzer.Common;
 
-namespace SonarAnalyzer.UnitTest.Helpers;
+namespace SonarAnalyzer.UnitTest.AnalysisContext;
 
 public partial class SonarAnalysisContextTest
 {
